+++ tests/PyroTests/test_socket.py
@@ -5,10 +5,12 @@ Pyro - Python Remote Objects.  Copyright by Irmen de Jong (irmen@razorvine.net).
 """
 
 import socket, os, sys
+import time
 import Pyro4.socketutil as SU
-from Pyro4 import threadutil
+from Pyro4 import threadutil, errors
 from Pyro4.socketserver.multiplexserver import SocketServer_Select, SocketServer_Poll
 from Pyro4.socketserver.threadpoolserver import SocketServer_Threadpool
+from Pyro4.core import MessageFactory, Daemon
 import Pyro4
 from testsupport import *
 
@@ -298,12 +300,28 @@ class TestSocketutil(unittest.TestCase):
         cs.close()
 
 class ServerCallback(object):
-    def handshake(self, connection):
+    def _handshake(self, connection):
         if not isinstance(connection, SU.SocketConnection):
             raise TypeError("handshake expected SocketConnection parameter")
+        msg=MessageFactory.createMessage(MessageFactory.MSG_CONNECTOK, None, 0, 1)
+        connection.send(msg)
+        return True
     def handleRequest(self, connection):
         if not isinstance(connection, SU.SocketConnection):
             raise TypeError("handleRequest expected SocketConnection parameter")
+        msgType, flags, seq, data = MessageFactory.getMessage(connection, [MessageFactory.MSG_PING])
+        if msgType == MessageFactory.MSG_PING:
+            msg = MessageFactory.createMessage(MessageFactory.MSG_PING, None, 0, 0)
+            connection.send(msg)
+        else:
+            print("unhandled message type", msgType)
+            connection.close()
+
+
+class ServerCallback_BrokenHandshake(ServerCallback):
+    def _handshake(self, connection):
+        raise ZeroDivisionError("handshake crashed (on purpose)")
+
 
 class TestSocketServer(unittest.TestCase):
     def testServer_thread(self):
@@ -354,6 +372,128 @@ class TestSocketServer(unittest.TestCase):
         serv.close()
         serv.close()
         self.assertTrue(serv.sock is None)
+
+
+
+
+@unittest.skipUnless(SU.hasSelect, "requires select()")
+class TestServerDOS_select(unittest.TestCase):
+
+    def setUp(self):
+        self.orig_poll_timeout = Pyro4.config.POLLTIMEOUT
+        self.orig_comm_timeout = Pyro4.config.COMMTIMEOUT
+        Pyro4.config.POLLTIMEOUT = 0.5
+        Pyro4.config.COMMTIMEOUT = 0.5
+        self.socket_server = SocketServer_Select
+
+    def tearDown(self):
+        Pyro4.config.POLLTIMEOUT = self.orig_poll_timeout
+        Pyro4.config.COMMTIMEOUT = self.orig_comm_timeout
+
+    class ServerThread(threadutil.Thread):
+        def __init__(self, server, daemon):
+            threadutil.Thread.__init__(self)
+            self.serv = server()
+            self.serv.init(daemon(), "localhost", 0)
+            self.locationStr = self.serv.locationStr
+            self.stop_loop = threadutil.Event()
+        def run(self):
+            self.serv.loop(loopCondition=lambda: not self.stop_loop.is_set())
+            self.serv.close()
+
+    def testConnectCrash(self):
+        serv_thread = TestServerDOS_select.ServerThread(self.socket_server, ServerCallback_BrokenHandshake)
+        serv_thread.start()
+        time.sleep(0.2)
+        self.assertTrue(serv_thread.is_alive(), "server thread failed to start")
+        try:
+            host, port = serv_thread.locationStr.split(':')
+            port = int(port)
+            try:
+                # first connection attempt (will fail because server daemon _handshake crashes)
+                csock = SU.createSocket(connect=(host, port))
+                conn = SU.SocketConnection(csock, "uri")
+                MessageFactory.getMessage(conn, MessageFactory.MSG_CONNECTOK)
+            except errors.ConnectionClosedError as x:
+                pass
+            conn.close()
+            try:
+                # second connection attempt, should still work (i.e. server should still be running)
+                csock = SU.createSocket(connect=(host, port))
+                conn = SU.SocketConnection(csock, "uri")
+                MessageFactory.getMessage(conn, MessageFactory.MSG_CONNECTOK)
+            except errors.ConnectionClosedError as x:
+                pass
+
+        finally:
+            conn.close()
+            serv_thread.stop_loop.set()
+            serv_thread.join()
+
+    def testInvalidMessageCrash(self):
+        serv_thread = TestServerDOS_select.ServerThread(self.socket_server, Daemon)
+        serv_thread.start()
+        time.sleep(0.2)
+        self.assertTrue(serv_thread.is_alive(), "server thread failed to start")
+
+        def connect(host, port):
+            # connect to the server
+            csock = SU.createSocket(connect=(host, port))
+            conn = SU.SocketConnection(csock, "uri")
+            # get the handshake/connect response
+            MessageFactory.getMessage(conn, [MessageFactory.MSG_CONNECTOK])
+            return conn
+
+        try:
+            host, port = serv_thread.locationStr.split(':')
+            port = int(port)
+            conn = connect(host, port)
+            # invoke something, but screw up the message (in this case, mess with the protocol version)
+            orig_protocol_version = Pyro4.constants.PROTOCOL_VERSION
+            Pyro4.constants.PROTOCOL_VERSION = 9999
+            msg = MessageFactory.createMessage(MessageFactory.MSG_PING, None, 0, 0)
+            Pyro4.constants.PROTOCOL_VERSION = orig_protocol_version
+            conn.send(msg)   # this should cause an error in the server because of invalid msg
+            try:
+                msgType, flags, seq, data = MessageFactory.getMessage(conn, [MessageFactory.MSG_RESULT])
+                self.assertTrue(b"Traceback" in data)
+                self.assertTrue(b"ProtocolError" in data)
+            except errors.ConnectionClosedError:
+                # invalid message can have caused the connection to be closed, this is fine
+                pass
+            # invoke something again, this should still work (server must still be running)
+            conn.close()
+            conn = connect(host, port)
+            msg = MessageFactory.createMessage(MessageFactory.MSG_PING, b"abc", 0, 999)
+            Pyro4.constants.PROTOCOL_VERSION = orig_protocol_version
+            conn.send(msg)
+            msgType, flags, seq, data = MessageFactory.getMessage(conn, [MessageFactory.MSG_PING])
+            self.assertEqual(MessageFactory.MSG_PING, msgType)
+            self.assertEqual(999, seq)
+            self.assertEqual(b"", data)
+        finally:
+            conn.close()
+            serv_thread.stop_loop.set()
+            serv_thread.join()
+
+
+@unittest.skipUnless(SU.hasPoll, "requires poll()")
+class TestServerDOS_poll(TestServerDOS_select):
+    def setUp(self):
+        super(TestServerDOS_poll, self).setUp()
+        self.socket_server = SocketServer_Poll
+
+class TestServerDOS_threading(TestServerDOS_select):
+    def setUp(self):
+        super(TestServerDOS_threading, self).setUp()
+        self.socket_server = SocketServer_Threadpool
+        self.orig_maxthreads = Pyro4.config.THREADPOOL_MAXTHREADS
+        Pyro4.config.THREADPOOL_MAXTHREADS = 1
+
+    def tearDown(self):
+        Pyro4.config.THREADPOOL_MAXTHREADS = self.orig_maxthreads
+
+
 
 
 if __name__ == "__main__":